+++ connector/solarwindsentityconnector/internal/event_detector_test.go
@@ -53,11 +53,11 @@ func TestDetectLog_EntityAndRelationshipEvents(t *testing.T) {
 		Attributes: []string{"attr"},
 	}
 	entityEvent := config.ParsedEntityEvent[ottllog.TransformContext]{
-		Definition:   &config.EntityEvent{Type: entity.Type},
+		Definition:   &config.EntityEvent{Type: entity.Type, Action: EventUpdateAction},
 		ConditionSeq: seq,
 	}
 	relationshipEvent := config.ParsedRelationshipEvent[ottllog.TransformContext]{
-		Definition:   &config.RelationshipEvent{Type: relationshipType, Source: entity.Type, Destination: entity.Type},
+		Definition:   &config.RelationshipEvent{Type: relationshipType, Source: entity.Type, Destination: entity.Type, Action: EventUpdateAction},
 		ConditionSeq: seq,
 	}
 	eventsGroup := config.EventsGroup[ottllog.TransformContext]{
@@ -117,7 +117,7 @@ func TestDetectLog_NoEvents(t *testing.T) {
 		Attributes: []string{"attr"},
 	}
 	entityEvent := config.ParsedEntityEvent[ottllog.TransformContext]{
-		Definition:   &config.EntityEvent{Type: entity.Type},
+		Definition:   &config.EntityEvent{Type: entity.Type, Action: EventUpdateAction},
 		ConditionSeq: seq,
 	}
 	eventsGroup := config.EventsGroup[ottllog.TransformContext]{
@@ -172,11 +172,11 @@ func TestDetectMetric_EntityAndRelationshipEvents(t *testing.T) {
 		Attributes: []string{"attr"},
 	}
 	entityEvent := config.ParsedEntityEvent[ottlmetric.TransformContext]{
-		Definition:   &config.EntityEvent{Type: entity.Type},
+		Definition:   &config.EntityEvent{Type: entity.Type, Action: EventUpdateAction},
 		ConditionSeq: seq,
 	}
 	relationshipEvent := config.ParsedRelationshipEvent[ottlmetric.TransformContext]{
-		Definition:   &config.RelationshipEvent{Type: relationshipType, Source: entity.Type, Destination: entity.Type},
+		Definition:   &config.RelationshipEvent{Type: relationshipType, Source: entity.Type, Destination: entity.Type, Action: EventUpdateAction},
 		ConditionSeq: seq,
 	}
 	eventsGroup := config.EventsGroup[ottlmetric.TransformContext]{
@@ -374,7 +374,7 @@ func TestCreateEntity(t *testing.T) {
 		config.EventsGroup[ottlmetric.TransformContext]{},
 		nil)
 
-	entityEvent, err := eventDetector.createEntity(resourceAttrs, &config.EntityEvent{Type: entity.Type})
+	entityEvent, err := eventDetector.createEntityEvent(resourceAttrs, &config.EntityEvent{Type: entity.Type, Action: "update"})
 	assert.Nil(t, err)
 	logs := plog.NewLogs()
 	logRecords := CreateEventLog(&logs)
@@ -415,7 +415,7 @@ func TestCreateEntityWithNoAttributes(t *testing.T) {
 		config.EventsGroup[ottllog.TransformContext]{},
 		config.EventsGroup[ottlmetric.TransformContext]{},
 		nil)
-	_, err := eventDetector.createEntity(resourceAttrs, &config.EntityEvent{Type: entity.Type})
+	_, err := eventDetector.createEntityEvent(resourceAttrs, &config.EntityEvent{Type: entity.Type})
 	assert.NotNil(t, err)
 }
 
@@ -441,6 +441,7 @@ func TestCreateRelationshipEvent(t *testing.T) {
 		Type:        "MemberOf",
 		Source:      srcEntity.Type,
 		Destination: destEntity.Type,
+		Action:      "update",
 	}
 
 	eventDetector := NewEventDetector(
@@ -454,7 +455,7 @@ func TestCreateRelationshipEvent(t *testing.T) {
 		config.EventsGroup[ottlmetric.TransformContext]{},
 		nil)
 
-	relationshipEvent, err := eventDetector.createRelationship(resourceAttrs, &relationship)
+	relationshipEvent, err := eventDetector.createRelationshipEvent(resourceAttrs, &relationship)
 	assert.Nil(t, err)
 	logs := plog.NewLogs()
 	logRecords := CreateEventLog(&logs)
@@ -507,7 +508,7 @@ func TestCreateRelationshipEventWithNoAttributes(t *testing.T) {
 		config.EventsGroup[ottlmetric.TransformContext]{},
 		nil)
 
-	_, err := eventDetector.createRelationship(resourceAttrs, &relationship)
+	_, err := eventDetector.createRelationshipEvent(resourceAttrs, &relationship)
 	assert.NotNil(t, err)
 }
 
@@ -541,7 +542,7 @@ func TestCreateRelationshipEventWithoutResourceAttributes(t *testing.T) {
 		config.EventsGroup[ottlmetric.TransformContext]{},
 		nil)
 
-	_, err := eventDetector.createRelationship(resourceAttrs, &relationship)
+	_, err := eventDetector.createRelationshipEvent(resourceAttrs, &relationship)
 	assert.NotNil(t, err)
 }
 
@@ -561,6 +562,7 @@ func TestCreateSameTypeRelationshipEvent(t *testing.T) {
 		Type:        "VirtualizationTopologyConnection",
 		Source:      "KubernetesCluster",
 		Destination: "KubernetesCluster",
+		Action:      "update",
 	}
 
 	eventDetector := NewEventDetector(
@@ -573,7 +575,7 @@ func TestCreateSameTypeRelationshipEvent(t *testing.T) {
 		config.EventsGroup[ottlmetric.TransformContext]{},
 		nil)
 
-	relationshipEvent, err := eventDetector.createRelationship(resourceAttrs, &relationship)
+	relationshipEvent, err := eventDetector.createRelationshipEvent(resourceAttrs, &relationship)
 	assert.Nil(t, err)
 	logs := plog.NewLogs()
 	logRecords := CreateEventLog(&logs)
@@ -620,7 +622,7 @@ func TestCreateSameTypeRelationshipEventWithNoAttributesSameType(t *testing.T) {
 		config.EventsGroup[ottlmetric.TransformContext]{},
 		nil)
 
-	_, err := eventDetector.createRelationship(resourceAttrs, &relationship)
+	_, err := eventDetector.createRelationshipEvent(resourceAttrs, &relationship)
 	assert.NotNil(t, err)
 }
 
@@ -648,7 +650,7 @@ func TestCreateSameTypeRelationshipEventWithoutResourceAttributes(t *testing.T) 
 		config.EventsGroup[ottlmetric.TransformContext]{},
 		nil)
 
-	_, err := eventDetector.createRelationship(resourceAttrs, &relationship)
+	_, err := eventDetector.createRelationshipEvent(resourceAttrs, &relationship)
 	assert.NotNil(t, err)
 }
 
@@ -669,7 +671,9 @@ func TestCollectEventsWithEntitiesWhenAttributesArePresent(t *testing.T) {
 		config.EventsGroup[ottllog.TransformContext]{},
 		config.EventsGroup[ottlmetric.TransformContext]{},
 		nil)
-	events, err := eventBuilder.collectEvents(resourceAttrs, []*config.EntityEvent{{Type: testEntity.Type}}, nil)
+	events, err := eventBuilder.collectEvents(resourceAttrs, []*config.EntityEvent{{
+		Type:   testEntity.Type,
+		Action: "update"}}, nil)
 	require.NoError(t, err)
 	require.Len(t, events, 1)
 	logs := plog.NewLogs()
@@ -732,7 +736,9 @@ func TestCollectEventsWithEntitiesWhenAttributeIsMissing(t *testing.T) {
 		config.EventsGroup[ottllog.TransformContext]{},
 		config.EventsGroup[ottlmetric.TransformContext]{},
 		nil)
-	events, err := eventBuilder.collectEvents(resourceAttrs, []*config.EntityEvent{{Type: testEntity.Type}}, nil)
+	events, err := eventBuilder.collectEvents(resourceAttrs, []*config.EntityEvent{{
+		Type:   testEntity.Type,
+		Action: "update"}}, nil)
 	require.NoError(t, err)
 	require.Len(t, events, 1)
 	logs := plog.NewLogs()
@@ -759,7 +765,10 @@ func TestCollectEventsWithRelationshipsWhenAttributesArePresent(t *testing.T) {
 	// arrange
 	srcEntity := config.Entity{Type: "KubernetesCluster", IDs: []string{"id1"}, Attributes: []string{"attr1"}}
 	destEntity := config.Entity{Type: "KubernetesNamespace", IDs: []string{"id2"}, Attributes: []string{"attr2"}}
-	testRelationship := config.RelationshipEvent{Source: srcEntity.Type, Destination: destEntity.Type}
+	testRelationship := config.RelationshipEvent{
+		Source:      srcEntity.Type,
+		Destination: destEntity.Type,
+		Action:      "update"}
 	resourceAttrs := pcommon.NewMap()
 	resourceAttrs.PutStr("id1", "idvalue1")
 	resourceAttrs.PutStr("id2", "idvalue2")
@@ -810,7 +819,10 @@ func TestCollectEventsWithRelationshipsWhenAttributesArePresent(t *testing.T) {
 func TestAppendSameTypeRelationshipUpdateEventWhenAttributesArePresent(t *testing.T) {
 	// arrange
 	entity := config.Entity{Type: "KubernetesCluster", IDs: []string{"id"}, Attributes: []string{"attr"}}
-	testRelationship := config.RelationshipEvent{Source: "KubernetesCluster", Destination: "KubernetesCluster"}
+	testRelationship := config.RelationshipEvent{
+		Source:      "KubernetesCluster",
+		Destination: "KubernetesCluster",
+		Action:      "update"}
 	resourceAttrs := pcommon.NewMap()
 	resourceAttrs.PutStr("src.id", "idvalue1")
 	resourceAttrs.PutStr("dst.id", "idvalue2")
@@ -916,7 +928,11 @@ func TestCollectEventsWithRelationshipAttribute(t *testing.T) {
 	// arrange
 	srcEntity := config.Entity{Type: "KubernetesCluster", IDs: []string{"id1"}}
 	destEntity := config.Entity{Type: "KubernetesNamespace", IDs: []string{"id2"}}
-	testRelationship := config.RelationshipEvent{Source: "KubernetesCluster", Destination: "KubernetesNamespace", Attributes: []string{"relationshipAttr"}}
+	testRelationship := config.RelationshipEvent{
+		Source:      "KubernetesCluster",
+		Destination: "KubernetesNamespace",
+		Attributes:  []string{"relationshipAttr"},
+		Action:      "update"}
 	resourceAttrs := pcommon.NewMap()
 	resourceAttrs.PutStr("id1", "idvalue1")
 	resourceAttrs.PutStr("id2", "idvalue2")
@@ -957,7 +973,11 @@ func TestCollectEventsWithRelationshipAttribute(t *testing.T) {
 func TestAppendSameTypeRelationshipUpdateEventWithRelationshipAttribute(t *testing.T) {
 	// arrange
 	entity := config.Entity{Type: "KubernetesCluster", IDs: []string{"id"}}
-	testRelationship := config.RelationshipEvent{Source: "KubernetesCluster", Destination: "KubernetesCluster", Attributes: []string{"relationshipAttr"}}
+	testRelationship := config.RelationshipEvent{
+		Source:      "KubernetesCluster",
+		Destination: "KubernetesCluster",
+		Attributes:  []string{"relationshipAttr"},
+		Action:      "update"}
 	resourceAttrs := pcommon.NewMap()
 	resourceAttrs.PutStr("src.id", "idvalue1")
 	resourceAttrs.PutStr("dst.id", "idvalue2")